Phase 1: Code Refactoring
